+++ components/analytics/actions/catalog-quadrant.tsx
@@ -2,32 +2,18 @@
 
 import { useMemo } from "react";
 import { CartesianGrid, Cell, ReferenceLine, Scatter, ScatterChart, Tooltip, XAxis, YAxis, ZAxis } from "recharts";
+import {
+  QUADRANT_COLORS,
+  QuadrantLegend,
+  type QuadrantPoint,
+  QuadrantTooltip
+} from "@/components/analytics/actions/catalog-quadrant-overlay";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import type { ActionLeaderboardRow } from "@/lib/analytics/actions-api";
 
 type Props = {
   rows: ActionLeaderboardRow[];
   minAdoption?: number;
-};
-
-type Point = {
-  id: string;
-  x: number;
-  y: number;
-  size: number;
-  title: string;
-  adoption_count: number;
-  completion_count: number;
-  completion_rate: number;
-  quadrant: "star" | "niche" | "popular_failing" | "kill" | "boundary";
-};
-
-const QUADRANT_COLORS: Record<Point["quadrant"], string> = {
-  star: "hsl(140, 60%, 40%)",
-  niche: "hsl(190, 55%, 45%)",
-  popular_failing: "hsl(35, 70%, 50%)",
-  kill: "hsl(0, 65%, 55%)",
-  boundary: "hsl(220, 8%, 55%)"
 };
 
 function median(values: number[]) {
@@ -40,7 +26,7 @@ function median(values: number[]) {
 export function CatalogQuadrant({ rows, minAdoption = 1 }: Props) {
   const { points, xMedian, yMedian, useLogScale } = useMemo(() => {
     const filtered = rows.filter((r) => r.adoption_count >= minAdoption);
-    if (filtered.length === 0) return { points: [] as Point[], xMedian: 0, yMedian: 0, useLogScale: false };
+    if (filtered.length === 0) return { points: [] as QuadrantPoint[], xMedian: 0, yMedian: 0, useLogScale: false };
 
     const xValues = filtered.map((r) => r.adoption_count);
     const xMin = Math.min(...xValues);
@@ -50,10 +36,10 @@ export function CatalogQuadrant({ rows, minAdoption = 1 }: Props) {
     const xMed = median(xValues);
     const yMed = median(filtered.map((r) => r.completion_rate));
 
-    const pts: Point[] = filtered.map((r) => {
+    const pts: QuadrantPoint[] = filtered.map((r) => {
       const aboveX = r.adoption_count > xMed;
       const aboveY = r.completion_rate > yMed;
-      let quadrant: Point["quadrant"];
+      let quadrant: QuadrantPoint["quadrant"];
       if (r.adoption_count === xMed || r.completion_rate === yMed) quadrant = "boundary";
       else if (aboveX && aboveY) quadrant = "star";
       else if (!aboveX && aboveY) quadrant = "niche";
@@ -150,27 +136,7 @@ export function CatalogQuadrant({ rows, minAdoption = 1 }: Props) {
               }}
             />
             <ReferenceLine y={yMedian} stroke="var(--muted-foreground)" strokeDasharray="3 3" strokeOpacity={0.4} />
-            <Tooltip
-              cursor={{ strokeDasharray: "3 3", stroke: "var(--border)" }}
-              content={({ active, payload }) => {
-                if (!active || !payload?.length) return null;
-                const point = payload[0].payload as Point;
-                return (
-                  <div className="rounded-md border bg-popover px-3 py-2 text-xs shadow-md">
-                    <div className="font-medium">{point.title}</div>
-                    <div className="text-muted-foreground">{quadrantLabel(point.quadrant)}</div>
-                    <div className="mt-1 grid grid-cols-2 gap-x-3 gap-y-0.5 text-muted-foreground">
-                      <span>Adoption</span>
-                      <span className="text-right text-foreground">{point.adoption_count}</span>
-                      <span>Completed</span>
-                      <span className="text-right text-foreground">{point.completion_count}</span>
-                      <span>Completion</span>
-                      <span className="text-right text-foreground">{(point.completion_rate * 100).toFixed(1)}%</span>
-                    </div>
-                  </div>
-                );
-              }}
-            />
+            <Tooltip cursor={{ strokeDasharray: "3 3", stroke: "var(--border)" }} content={QuadrantTooltip} />
             <Scatter data={points} fillOpacity={0.75}>
               {points.map((point) => (
                 <Cell key={point.id} fill={QUADRANT_COLORS[point.quadrant]} />
@@ -178,31 +144,9 @@ export function CatalogQuadrant({ rows, minAdoption = 1 }: Props) {
             </Scatter>
           </ScatterChart>
 
-          <div className="pointer-events-none absolute inset-0 grid grid-cols-2 grid-rows-2 p-3 text-[10px] uppercase tracking-wide">
-            <div className="self-start justify-self-start text-muted-foreground/70">Niche but effective</div>
-            <div className="self-start justify-self-end text-emerald-600/80 dark:text-emerald-400/80">Stars</div>
-            <div className="self-end justify-self-start text-destructive/70">Kill candidates</div>
-            <div className="self-end justify-self-end text-amber-600/80 dark:text-amber-400/80">
-              Popular but failing
-            </div>
-          </div>
+          <QuadrantLegend />
         </div>
       </CardContent>
     </Card>
   );
-}
-
-function quadrantLabel(q: Point["quadrant"]) {
-  switch (q) {
-    case "star":
-      return "Star — high adoption, high completion";
-    case "niche":
-      return "Niche but effective";
-    case "popular_failing":
-      return "Popular but failing";
-    case "kill":
-      return "Kill candidate";
-    default:
-      return "On the median";
-  }
 }
