+++ components/analytics/actions/forest-plot.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { useMemo } from "react";
+import { ForestPlotRow, formatPct } from "@/components/analytics/actions/forest-plot-row";
 import type { ActionCorrelationFactor } from "@/lib/analytics/actions-api";
 
 type Props = {
@@ -16,12 +17,6 @@ const MIN_DOT_RADIUS = 3;
 const MAX_DOT_RADIUS = 7;
 
 const TICKS = [-0.5, -0.25, 0, 0.25, 0.5];
-
-function formatPct(value: number | null): string {
-  if (value === null) return "—";
-  const sign = value > 0 ? "+" : "";
-  return `${sign}${(value * 100).toFixed(1)}%`;
-}
 
 export function ForestPlot({ factors }: Props) {
   const width = 720;
@@ -113,97 +108,18 @@ export function ForestPlot({ factors }: Props) {
         ))}
 
         {/* Rows */}
-        {factors.map((factor, i) => {
-          const y = TOP_PADDING + ROW_HEIGHT * (i + 0.5);
-          const isInsufficient = factor.insufficient_data || factor.lift === null;
-          const lift = factor.lift ?? 0;
-          const lower = factor.ci_low ?? lift;
-          const upper = factor.ci_high ?? lift;
-
-          const color = isInsufficient
-            ? "var(--muted-foreground)"
-            : lift > 0
-              ? "hsl(140, 60%, 40%)"
-              : lift < 0
-                ? "hsl(0, 65%, 55%)"
-                : "var(--muted-foreground)";
-
-          return (
-            <g key={factor.factor}>
-              <text x={8} y={y} dy="0.32em" fontSize={12} fill="var(--foreground)">
-                {factor.label}
-              </text>
-
-              {isInsufficient ? (
-                <text
-                  x={(plotLeft + plotRight) / 2}
-                  y={y}
-                  dy="0.32em"
-                  textAnchor="middle"
-                  fontSize={11}
-                  fill="var(--muted-foreground)"
-                >
-                  Not enough data
-                </text>
-              ) : (
-                <>
-                  <line
-                    x1={xScale(lower)}
-                    x2={xScale(upper)}
-                    y1={y}
-                    y2={y}
-                    stroke={color}
-                    strokeOpacity={0.5}
-                    strokeWidth={2}
-                  />
-                  <line
-                    x1={xScale(lower)}
-                    x2={xScale(lower)}
-                    y1={y - 4}
-                    y2={y + 4}
-                    stroke={color}
-                    strokeOpacity={0.5}
-                    strokeWidth={2}
-                  />
-                  <line
-                    x1={xScale(upper)}
-                    x2={xScale(upper)}
-                    y1={y - 4}
-                    y2={y + 4}
-                    stroke={color}
-                    strokeOpacity={0.5}
-                    strokeWidth={2}
-                  />
-                  <circle
-                    cx={xScale(lift)}
-                    cy={y}
-                    r={dotRadius(Math.min(factor.with.n, factor.without.n))}
-                    fill={color}
-                  >
-                    <title>
-                      {`${factor.label}: lift ${formatPct(lift)} (95% CI ${formatPct(lower)} … ${formatPct(upper)})\n` +
-                        `with: ${factor.with.n} samples, ${(factor.with.completion_rate * 100).toFixed(1)}% completed\n` +
-                        `without: ${factor.without.n} samples, ${(factor.without.completion_rate * 100).toFixed(1)}% completed`}
-                    </title>
-                  </circle>
-                  <text
-                    x={xScale(lift)}
-                    y={y - dotRadius(Math.min(factor.with.n, factor.without.n)) - 4}
-                    fontSize={10}
-                    fill="var(--muted-foreground)"
-                    textAnchor="middle"
-                  >
-                    {formatPct(lift)}
-                  </text>
-                </>
-              )}
-
-              <text x={rightStrip} y={y} dy="0.32em" fontSize={11} fill="var(--muted-foreground)" textAnchor="end">
-                {factor.with.n} / {factor.without.n}
-              </text>
-            </g>
-          );
-        })}
+        {factors.map((factor, i) => (
+          <ForestPlotRow
+            key={factor.factor}
+            factor={factor}
+            y={TOP_PADDING + ROW_HEIGHT * (i + 0.5)}
+            plotLeft={plotLeft}
+            plotRight={plotRight}
+            rightStrip={rightStrip}
+            xScale={xScale}
+            dotRadius={dotRadius}
+          />
+        ))}
       </svg>
     </div>
   );
