+++ components/analytics/insights/cohort-builder.tsx
@@ -2,15 +2,13 @@
 
 import { Plus, RotateCcw } from "lucide-react";
 import { Fragment, useState } from "react";
+import { CohortOrgExtras } from "@/components/analytics/insights/cohort-org-extras";
 import { FactFilterRow } from "@/components/analytics/insights/fact-filter-row";
 import { MultiCheckGroup } from "@/components/analytics/insights/multi-check-group";
 import { TagPicker } from "@/components/analytics/insights/tag-picker";
 import { WorkshopPicker } from "@/components/analytics/insights/workshop-picker";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { useCohort } from "@/lib/analytics/insights/cohort-context";
 import type { FactFilter, OrgFilters } from "@/lib/analytics/insights/cohort-spec";
 import { useLiveDataset } from "@/lib/analytics/use-live-dataset";
@@ -52,19 +50,6 @@ export function CohortBuilder({ embedded = false }: Props = {}) {
     updateFactFilters([...spec.fact_filters, { key: "", values: [] }]);
   }
 
-  const trackedActionsValue =
-    spec.org_filters.has_tracked_actions === true
-      ? "yes"
-      : spec.org_filters.has_tracked_actions === false
-        ? "no"
-        : "any";
-
-  function setTrackedActions(value: string) {
-    if (value === "yes") updateOrg({ has_tracked_actions: true });
-    else if (value === "no") updateOrg({ has_tracked_actions: false });
-    else updateOrg({ has_tracked_actions: undefined });
-  }
-
   const body = (
     <Fragment>
       <div className="grid gap-5 md:grid-cols-2">
@@ -100,46 +85,7 @@ export function CohortBuilder({ embedded = false }: Props = {}) {
         onChange={(next) => updateOrg({ workshop_uuids: next })}
       />
 
-      <div className="grid gap-5 md:grid-cols-2">
-        <div className="space-y-2">
-          <span className="text-xs font-medium uppercase tracking-wide text-muted-foreground">Signed up</span>
-          <div className="flex items-center gap-2">
-            <div className="flex flex-col gap-1">
-              <Label className="text-xs">From</Label>
-              <Input
-                type="date"
-                value={spec.org_filters.signed_up_from?.slice(0, 10) ?? ""}
-                onChange={(e) => updateOrg({ signed_up_from: e.target.value || undefined })}
-                className="h-8 text-sm"
-              />
-            </div>
-            <div className="flex flex-col gap-1">
-              <Label className="text-xs">To</Label>
-              <Input
-                type="date"
-                value={spec.org_filters.signed_up_to?.slice(0, 10) ?? ""}
-                onChange={(e) => updateOrg({ signed_up_to: e.target.value || undefined })}
-                className="h-8 text-sm"
-              />
-            </div>
-          </div>
-        </div>
-
-        <div className="space-y-2">
-          <span className="text-xs font-medium uppercase tracking-wide text-muted-foreground">Has tracked actions</span>
-          <ToggleGroup
-            type="single"
-            value={trackedActionsValue}
-            onValueChange={(v) => v && setTrackedActions(v)}
-            variant="outline"
-            size="sm"
-          >
-            <ToggleGroupItem value="any">Any</ToggleGroupItem>
-            <ToggleGroupItem value="yes">With actions</ToggleGroupItem>
-            <ToggleGroupItem value="no">Without</ToggleGroupItem>
-          </ToggleGroup>
-        </div>
-      </div>
+      <CohortOrgExtras orgFilters={spec.org_filters} onChange={updateOrg} />
 
       <div className="space-y-2">
         <div className="flex items-center justify-between">
